requirements.txt
# PrepMe RAG package
chromadb>=0.5.0
sentence-transformers>=3.0.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
pypdf>=4.0.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
python-dotenv>=1.0.0
openai>=1.30.0
tqdm>=4.66.0
